+++ lib/qc/modules/pre_dispatch/resources/pre_dispatch_lot_list_request.g.dart
@@ -6,14 +6,18 @@ part of 'pre_dispatch_lot_list_request.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-PreDispatchLotRequest _$PreDispatchLotRequestFromJson(Map<String, dynamic> json) => PreDispatchLotRequest(
+PreDispatchLotRequest _$PreDispatchLotRequestFromJson(
+        Map<String, dynamic> json) =>
+    PreDispatchLotRequest(
       pageNo: (json['offset'] as num?)?.toInt(),
       pageSize: (json['pageSize'] as num?)?.toInt(),
-      filterMap:
-          json['filterObjectMap'] == null ? null : FilterMap.fromJson(json['filterObjectMap'] as Map<String, dynamic>),
+      filterMap: json['filterObjectMap'] == null
+          ? null
+          : FilterMap.fromJson(json['filterObjectMap'] as Map<String, dynamic>),
     );
 
-Map<String, dynamic> _$PreDispatchLotRequestToJson(PreDispatchLotRequest instance) {
+Map<String, dynamic> _$PreDispatchLotRequestToJson(
+    PreDispatchLotRequest instance) {
   final val = <String, dynamic>{};
 
   void writeNotNull(String key, dynamic value) {
@@ -30,7 +34,9 @@ Map<String, dynamic> _$PreDispatchLotRequestToJson(PreDispatchLotRequest instanc
 
 FilterMap _$FilterMapFromJson(Map<String, dynamic> json) => FilterMap(
       searchQuery: json['q'] as String?,
-      lotType: (json['lt'] as List<dynamic>?)?.map((e) => (e as num).toInt()).toList(),
+      lotType: (json['lt'] as List<dynamic>?)
+          ?.map((e) => (e as num).toInt())
+          .toList(),
       barcode: json['br'] as String?,
     );