+++ lib/src/common/resources/lot_list_request.g.dart
@@ -6,11 +6,13 @@ part of 'lot_list_request.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-LotListRequest _$LotListRequestFromJson(Map<String, dynamic> json) => LotListRequest(
+LotListRequest _$LotListRequestFromJson(Map<String, dynamic> json) =>
+    LotListRequest(
       pageNo: (json['offset'] as num?)?.toInt(),
       pageSize: (json['pageSize'] as num?)?.toInt(),
-      filterMap:
-          json['filterObjectMap'] == null ? null : FilterMap.fromJson(json['filterObjectMap'] as Map<String, dynamic>),
+      filterMap: json['filterObjectMap'] == null
+          ? null
+          : FilterMap.fromJson(json['filterObjectMap'] as Map<String, dynamic>),
     );
 
 Map<String, dynamic> _$LotListRequestToJson(LotListRequest instance) {
@@ -30,8 +32,10 @@ Map<String, dynamic> _$LotListRequestToJson(LotListRequest instance) {
 
 FilterMap _$FilterMapFromJson(Map<String, dynamic> json) => FilterMap(
       searchQuery: json['q'] as String?,
-      lotType: (json['lt'] as List<dynamic>?)?.map((e) => (e as num).toInt()).toList(),
-      isStoreOut: (json['iso'] as num).toInt(),
+      lotType: (json['lt'] as List<dynamic>?)
+          ?.map((e) => (e as num).toInt())
+          .toList(),
+      isStoreOut: (json['iso'] as num?)?.toInt(),
     );
 
 Map<String, dynamic> _$FilterMapToJson(FilterMap instance) {